move other types to types folder, #18
