+++ components/common/types/card.types.ts
