+++ components/common/utils/getDailyItem.ts
@@ -1,4 +1,4 @@
-import { CardInterface } from "@/components/Card/card.types";
+import { CardInterface } from "@/components/common/types/card.types";
 
 export const getDailyItem = (items: CardInterface[]): CardInterface => {
   const today = new Date();
